Introducing bullet related systems, update version and packages
make a simple spawn, move and lifetime system for project tiles, update enum to support and prepare for each type of bullet to have different movement, upgrade new version, add visual effect graph package to prepare to make trail renderer for bullets instead of using trail renderer in unity because dots does not support this component

+++ Assets/AssetInventory/Editor/Scripts/GUI/SubViews/Metadata/MetadataUI.cs
@@ -64,7 +64,10 @@ namespace AssetInventory
                         }
                         if (GUILayout.Button(EditorGUIUtility.IconContent("TreeEditor.Trash", "|Remove metadata completely"), GUILayout.Width(30)))
                         {
-                            Metadata.DeleteDefinition(meta);
+                            if (EditorUtility.DisplayDialog("Delete Metadata Definitions", "Are you sure you want to delete this metadata definitions and all connected data? This action cannot be undone.", "Delete", "Cancel"))
+                            {
+                                Metadata.DeleteDefinition(meta);
+                            }
                         }
                         GUILayout.EndHorizontal();
                     }
@@ -81,7 +84,10 @@ namespace AssetInventory
             EditorGUI.BeginDisabledGroup(_metas == null || _metas.Count == 0);
             if (GUILayout.Button("Delete All"))
             {
-                _metas?.ForEach(Metadata.DeleteDefinition);
+                if (EditorUtility.DisplayDialog("Delete All Metadata Definitions", "Are you sure you want to delete all metadata definitions? This action cannot be undone.", "Delete", "Cancel"))
+                {
+                    _metas?.ForEach(Metadata.DeleteDefinition);
+                }
             }
             EditorGUI.EndDisabledGroup();
             GUILayout.EndHorizontal();
